Constructors

MultistageTrueAndErrorModels.MultistageTrueErrorModel — Type

struct MultistageTrueErrorModel{T<:Real} <: TrueAndErrorModels.AbstractTrueErrorModel{T<:Real}

A True and Error Model of backward vs.forward induction strategy in a two stage sequential decision making task (Deng et al., 2026).

Fields

b::T: the probability of using backward induction

sb::T: the probability of continuing a backward induction strategy in blocks 3 and 4

sf::T: the probability of continuing a forward induction strategy in blocks 3 and 4

ϵ::AbstractVector{T}: a vector of error probabilities with elements ϵ = [ϵ₁, ϵ₂, ϵ₃, ϵ₄] where ϵᵢ ≤ .50 and i indexes the block id.

Example

using MultistageTrueAndErrorModels

model = MultistageTrueErrorModel(;
    b = 0.1,
    sb = 0.8,
    sf = 0.4,
    ϵ = [0.1, 0.04, 0.02, 0.03]
)

data = rand(model, 200)
logpdf(model, data)

Documentation

Full documentation can be found at https://itsdfish.github.io/MultistageTrueAndErrorModels.jl/dev/

References

Birnbaum, M. H., & Quispe-Torreblanca, E. G. (2018). TEMAP2. R: True and error model analysis program in R. Judgment and Decision Making, 13(5), 428-440.

Deng, W., Kellen, D., & Hotaling, J. M. (2026). Toward the cognitive modeling of dynamic decision making. Psychonomic Bulletin & Review, 33(4), 127.

Lee, M. D. (2018). Bayesian methods for analyzing true-and-error models. Judgment and Decision Making, 13(6), 622-635.

Core Functions

TrueAndErrorModels.compute_probs — Function

compute_probs(dist::MultistageTrueErrorModel{T}) -> Any

Computes the joint probability for all 16 response categories.

Arguments

dist::MultistageTrueErrorModel{T}: a distribution object for a True and Error Model of multistage decision making.

Output

θ::Vector{T}: vector of joint response probabilities for the following response patterns across the four blocks:

BBBB

FBBB

BFBB

FFBB

BBFB

FBFB

BFFB

FFFB

BBBF

FBBF

BFBF

FFBF

BBFF

FBFF

BFFF

FFFF

where B corresponds to a response pattern consistent with backward induction, F corresponds to a response pattern consistent with forward induction, and the position of each element corresponds to block ID.

Example

using MultistageTrueAndErrorModels

model = MultistageTrueErrorModel(;
    b = 0.3,
    sb = 0.4,
    sf = 0.2,
    ϵ = [0.05, 0.06, 0.07, 0.08]
)
probs = compute_probs(model)

Utilities

TrueAndErrorModels.get_response_labels — Function

get_response_labels(::MultistageTrueErrorModel)

Returns a vector of response pattern labels.